requirements.txt
python-binance==1.0.19
pandas>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1